server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
passlib==1.7.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
psycopg2==2.9.10
tortoise-orm==0.23.0
databases==0.9.0
asyncpg==0.30.0
PyJWT==2.9.0